revert redis config name move redis uri to separate class
