+++ redis/src/test/scala/zio/redis/ClusterSpec.scala
@@ -1,5 +1,6 @@
 package zio.redis
 
+import zio.redis.executor.cluster.RedisUri
 import zio.test.Assertion._
 import zio.test._
